second HT block coder: support lossless encode

+++ src/lib/jp2/t1/part15/coding/coding_units.cpp
@@ -43,7 +43,7 @@
  *******************************************************************************/
 
 j2k_codeblock::j2k_codeblock(const uint32_t &idx, uint8_t orientation, uint8_t M_b, uint8_t R_b,
-                             uint8_t transformation, float stepsize, uint32_t band_stride, int16_t *ibuf,
+                             uint8_t transformation, float stepsize, uint32_t band_stride, int32_t *ibuf,
                              float *fbuf, uint32_t offset, const uint16_t &numlayers,
                              const uint8_t &codeblock_style, const element_siz &p0, const element_siz &p1,
                              const element_siz &s)
